Add files via upload
adding files saved down from the sagemaker notebook instances for the students. Mostly it is the fall 2021 project on detecting botnets from disqus user data. The data itself was too large to include in this repository and was pulled down from kaggle. If it is available somewhere else online though a cleaned up version of it should work with these notebooks.